+++ trackeval/datasets/_base_dataset.py
@@ -1,3 +1,5 @@
+from typing import List, Dict, Union
+
 import csv
 import io
 import os
@@ -9,22 +11,21 @@ from copy import deepcopy
 
 import numpy as np
 
-from .. import _timing
-from ..utils import TrackEvalException
+from trackeval import _timing
+from trackeval.exception import TrackEvalException
+from trackeval.types import FMT
 
 
 class _BaseDataset(ABC):
     @abstractmethod
     def __init__(self):
-        self.tracker_list = None
-        self.seq_list = None
-        self.class_list = None
+        self.tracker_list: List[str] = []
+        self.seq_list: List[str] = []
+        self.class_list: List[str] = []
         self.output_fol = None
         self.output_sub_fol = None
         self.should_classes_combine = True
         self.use_super_categories = False
-
-    # Functions to implement:
 
     @staticmethod
     @abstractmethod
@@ -47,10 +48,10 @@ class _BaseDataset(ABC):
     # Helper functions for all datasets:
 
     @classmethod
-    def get_class_name(cls):
+    def get_class_name(cls) -> str:
         return cls.__name__
 
-    def get_name(self):
+    def get_name(self) -> str:
         return self.get_class_name()
 
     def get_output_fol(self, tracker: str) -> str:
@@ -67,9 +68,9 @@ class _BaseDataset(ABC):
         return self.tracker_list, self.seq_list, self.class_list
 
     @_timing.time
-    def get_raw_seq_data(self, tracker, seq):
+    def get_raw_seq_data(self, tracker, seq) -> Dict[str, Union[int, FMT]]:
         """ Loads raw data (tracker and ground-truth) for a single tracker on a single sequence.
-        Raw data includes all of the information needed for both preprocessing and evaluation, for all classes.
+        Raw data includes all the information needed for both preprocessing and evaluation, for all classes.
         A later function (get_processed_seq_data) will perform such preprocessing and extract relevant information for
         the evaluation of each class.
 
@@ -248,7 +249,7 @@ class _BaseDataset(ABC):
         return ious
 
     @staticmethod
-    def _calculate_box_ious(bboxes1, bboxes2, box_format='xywh', do_ioa=False):
+    def _calculate_box_ious(bboxes1: FMT, bboxes2: FMT, box_format: str = 'xywh', do_ioa: bool = False) -> FMT:
         """ Calculates the IOU (intersection over union) between two arrays of boxes.
         Allows variable box formats ('xywh' and 'x0y0x1y1').
         If do_ioa (intersection over area) , then calculates the intersection over the area of boxes1 - this is commonly