+++ trackeval/metrics/__init__.py
@@ -1,3 +1,5 @@
+from typing import Union
+
 from .clear import CLEAR
 from .count import Count
 from trackeval.metrics.hota.hota import HOTA
@@ -6,3 +8,5 @@ from .ideucl import IDEucl
 from .j_and_f import JAndF
 from .track_map import TrackMAP
 from .vace import VACE
+
+MET = Union[HOTA, CLEAR, Identity]