+++ trackeval/metrics/_base_metric.py
@@ -4,7 +4,7 @@ from abc import ABC, abstractmethod
 import numpy as np
 
 from .. import _timing
-from ..utils import TrackEvalException
+from ..exception import TrackEvalException
 
 
 class _BaseMetric(ABC):