+++ trackeval/metrics/j_and_f.py
@@ -6,7 +6,7 @@ import numpy as np
 from scipy.optimize import linear_sum_assignment
 
 from .. import _timing
-from ..utils import TrackEvalException
+from ..exception import TrackEvalException
 from ._base_metric import _BaseMetric